+++ code/cell.modules/pkg.sys/net/src/web.ui/Label.Network/index.ts
@@ -1,0 +1,1 @@
+export * from './Label.Network';
