+++ code/cell.modules/pkg.sys/net/src/web.ui/Label/index.ts
@@ -1,3 +1,1 @@
 export * from './Label';
-export * from './Label.Peer';
-export * from './Label.Network';
